+++ doc/version_history.md
@@ -1,5 +1,14 @@
 ## 版本更新记录
 
+##### release
+
+- Global
+    - [Upgrade] 优化中间件脚本[docker-compose-base.yaml](../dockerfiles/docker-compose-base.yaml)
+- oauth-server
+    - [Upgrade] 优化 UserPasswordAuthenticationToken 序列化
+    - [Upgrade] token查询过滤过期的token
+    - [Upgrade] 调整token相关controller名称
+
 ##### v5.1.3
 
 - Global